import CiaoDE-1.15.0-15080.tar.gz

+++ ciao/contrib/hashtable/recycle.c
@@ -84,8 +84,7 @@ char   *purpose;
   char *x = (char *)ht_alloc(len);
   if (!x)
   {
-    fprintf(stderr, "malloc of %d failed for %s\n", 
-	    len, purpose);
+    fprintf(stderr, "malloc of %lu failed for %s\n", (unsigned long)len, purpose);
     exit(SUCCESS);
   }
   return x;
